pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "routa-fitness"
version = "0.1.4"
description = "Routa guardrails for architecture fitness, change-aware checks, and human review triggers"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Phodal" },
]
keywords = ["fitness-functions", "architecture", "quality-gates", "code-review", "ci"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "pyyaml>=6.0,<7",
    "tree-sitter-language-pack>=0.13,<1",
]

[project.urls]
Homepage = "https://github.com/phodal/entrix"
Repository = "https://github.com/phodal/entrix"
Issues = "https://github.com/phodal/entrix/issues"

[project.optional-dependencies]
graph = [
    "code-review-graph>=1.8,<2",
]
mcp = [
    "fastmcp>=0.1.0,<2",
]
dev = [
    "pytest>=8.0,<9",
    "ruff>=0.3.0,<1",
]

[project.scripts]
routa-fitness = "routa_fitness.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["routa_fitness"]

[tool.ruff]
line-length = 100
target-version = "py310"
